requirements.txt
python-telegram-bot==20.3
google-api-python-client==2.86.0
python-dotenv==1.0.0
aiohttp==3.8.5
asyncio==3.4.3